requirements.txt
discord.py==1.7.1
requests==2.25.1
beautifulsoup4==4.9.3
tabulate==0.8.9
mysql-connector==2.2.9
statcord.py==3.1.0
pytz==2021.1